requirements.txt
Pillow>=10.0.0
pillow-heif>=0.13.0
Flask>=3.0.0
gunicorn>=21.2.0

